+++ packages/components/src/components/bal-tooltip/test/bal-tooltip.visual.html
@@ -21,23 +21,24 @@
           <bal-button data-test="arrow-trigger" bal-tooltip="arrow-tooltip">Arrow</bal-button>
           <bal-button data-test="placement-right-trigger" bal-tooltip="placement-right-tooltip">Right</bal-button>
         </bal-stack>
-        
+
         <bal-tooltip id="tooltip-basic" data-test="basic" label="Basic Tooltip" reference="basic-tooltip">
-          <p>
-            Basic Tooltip
-          </p>
+          <p>Basic Tooltip</p>
         </bal-tooltip>
 
         <bal-tooltip id="tooltip-arrow" data-test="arrow" label="Arrow Tooltip" reference="arrow-tooltip" arrow>
-          <p>
-            Tooltip with arrow
-          </p>
+          <p>Tooltip with arrow</p>
         </bal-tooltip>
 
-        <bal-tooltip id="tooltip-placement-right" data-test="placement-right" label="Placement Right" reference="placement-right-tooltip" arrow placement="right">
-          <p>
-            Tooltip right placement
-          </p>
+        <bal-tooltip
+          id="tooltip-placement-right"
+          data-test="placement-right"
+          label="Placement Right"
+          reference="placement-right-tooltip"
+          arrow
+          placement="right"
+        >
+          <p>Tooltip right placement</p>
         </bal-tooltip>
       </main>
     </bal-doc-app>
